add image handling

+++ backend/src/logic/vacations-logic.ts
@@ -2,13 +2,15 @@ import { OkPacket } from "mysql";
 import { ResourceNotFoundErrorModel, ValidationErrorModel } from "../models/error-model";
 import VacationModel from "../models/vacation-model";
 import dal from "../utils/dal";
+import { v4 as uuid } from "uuid"; // import v4 function and rename to uuid
+import { promises } from "fs";
 
 // Function to get all vacations
 async function getAllVacations():Promise<VacationModel[]> {
     // Create SQL query - selects all columns from vacations table, with dates formatted
     const sqlQuery = `
         SELECT vacationID, destination, description, DATE_FORMAT(startDate, "%Y-%m-%d") as "startDate", DATE_FORMAT(endDate, "%Y-%m-%d") as "endDate", price, imageName
-        FROM vacations`;
+        FROM vacations;`;
 
     // Execute SQL query and save in variable to be returned
     const vacations = await dal.execute(sqlQuery);
@@ -36,10 +38,21 @@ async function getVacationById(id:number):Promise<VacationModel> {
 // Function to add a new vacation
 async function addVacation(vacation:VacationModel):Promise<VacationModel> {
     // Validate the vacation passed to function
-    const error = vacation.validate();
+    const errors = vacation.validate();
 
     // If there were errors in validation, throw an error (and quit function)
-    if (error) throw new ValidationErrorModel(error);
+    if (errors) throw new ValidationErrorModel(errors);
+
+    // Check if image was provided
+    if (vacation.image) {
+        // Extract file extension from image (for example ".jpg")
+        const extension = vacation.image.name.substring(vacation.image.name.lastIndexOf("."));
+        // Set image name as universal unique identifier (uuid)
+        vacation.imageName = uuid() + extension;
+        // Move image to server and name with uuid
+        await vacation.image.mv("../frontend/public/" + vacation.imageName);
+    }
+    else throw new ValidationErrorModel("No image provided!");
 
     // Create SQL query - add the information from the vacation passed to the function into the vacations table
     const sqlQuery = `
@@ -62,6 +75,47 @@ async function updateVacation(vacation:VacationModel):Promise<VacationModel> {
 
     // If there were errors in validation, throw an error (and quit function)
     if (error) throw new ValidationErrorModel(error);
+
+    // Create SQL query to get image name for vacation to update
+    const imgSqlQuery = `
+    SELECT imageName
+    FROM vacations
+    WHERE vacationID = ${vacation.vacationID};
+    `;
+
+    // Execute image SQL query and save the info
+    const imgInfo = await dal.execute(imgSqlQuery);
+    
+    // Save the image path
+    const imgPath = imgInfo?.[0]?.imageName;
+
+    // Get existing image path
+    const prevImgPath = "../frontend/public/" + imgPath;
+
+    // Create boolean flag to check if img path exists with default as true
+    let doesImgExist = true;
+    // Check that existing path exists
+    try {
+        await promises.access(prevImgPath);
+    } catch (error) {
+        // If existing path doesn't exist, set boolean flag as false
+        doesImgExist = false;
+    }
+    
+    // Check if image was provided
+    if (vacation.image) {        
+        // Delete existing path if exists (so can be overwritten)
+        if (doesImgExist) await promises.unlink(prevImgPath);
+
+        // Extract file extension from image (for example ".jpg")
+        const extension = vacation.image.name.substring(vacation.image.name.lastIndexOf("."));
+        // Set image name as universal unique identifier (uuid)
+        vacation.imageName = uuid() + extension;
+        // Move image to server and name with uuid
+        await vacation.image.mv("../frontend/public/" + vacation.imageName);
+    }
+    // If no image and no imageName provided, throw an error
+    else if (!vacation.imageName) throw new ValidationErrorModel("No image provided!");
 
     // Create SQL query - update the vacation based on information from the vacation passed to the function
     const sqlQuery = `
@@ -86,6 +140,19 @@ async function updateVacation(vacation:VacationModel):Promise<VacationModel> {
 }
 
 async function deleteVacation(id:number):Promise<void> {
+    // Create SQL query - selects all columns from vacations table for specific id, with dates formatted
+    const imgSqlQuery = `
+    SELECT imageName
+    FROM vacations
+    WHERE vacationID = ${id};
+    `;
+
+    // Execute SQL query and save the info
+    const imgInfo = await dal.execute(imgSqlQuery);
+
+    // If there isn't data for id being checked to delete, throw error 
+    if(!imgInfo[0]) throw new ResourceNotFoundErrorModel(id);
+    
     // Create SQL query - delete vacation from table based on id passed
     const sqlQuery = `
     DELETE FROM vacations
@@ -96,7 +163,13 @@ async function deleteVacation(id:number):Promise<void> {
     const info:OkPacket = await dal.execute(sqlQuery);
 
     // Check if ID exists by checking affectedRows of info, if it doesn't throw an error
-    if (info.affectedRows === 0) throw new ResourceNotFoundErrorModel(id);
+    // Commenting out as unneccesary in this instance as was checked earlier
+    // if (info.affectedRows <= 0) throw new ResourceNotFoundErrorModel(id);
+
+    // Get image path of deleted vacation
+    const imgPath = "../frontend/public/" + imgInfo?.[0]?.imageName;
+    // Delete image path
+    await promises.unlink(imgPath);
 }
 
 export default {
